:recycle: refactor: 체형 분석 및 결과 페이지 컴포넌트 분리
