+++ src/shared/types/body-type.ts
@@ -1,0 +1,9 @@
+export interface Question {
+  id: number
+  question: string
+  options: {
+    text: string
+    type: "natural" | "wave" | "straight"
+    image?: string
+  }[]
+}
